[mrc-v3] Add double coordinates
This adds dcrd[0-3] member objects to crd holding double precision versions
of the coordinates.

Coords are now generated in double and cast down to single to fill the
standard coord structs. Everything seems to work properly..

NB: very few io systems will actually write or read the double coords, so
watch out for that.

+++ include/mrc_crds_gen.h
@@ -6,7 +6,7 @@
 
 MRC_CLASS_DECLARE(mrc_crds_gen, struct mrc_crds_gen);
 
-void mrc_crds_gen_run(struct mrc_crds_gen *gen, float *xx, float *dx);
+void mrc_crds_gen_run(struct mrc_crds_gen *gen, double *xx, double *dx);
 
 #endif
